Refactor and simplify all CI job executor scripts
