+++ build-system/pr-check/checks.js
@@ -16,104 +16,84 @@
 'use strict';
 
 /**
- * @fileoverview
- * This script performs quick checks on the source code
- * prior to running unit and integration tests.
- * This is run during the CI stage = build; job = checks.
+ * @fileoverview Script that runs various checks during CI.
  */
 
-const {
-  abortTimedJob,
-  printChangeSummary,
-  startTimer,
-  stopTimer,
-  timedExecOrDie,
-} = require('./utils');
 const {determineBuildTargets} = require('./build-targets');
-const {isPullRequestBuild} = require('../common/ci');
 const {reportAllExpectedTests} = require('../tasks/report-test-status');
-const {runNpmChecks} = require('./npm-checks');
-const {setLoggingPrefix} = require('../common/logging');
+const {runCiJob} = require('./ci-job');
+const {timedExecOrDie} = require('./utils');
 
 const jobName = 'checks.js';
 
-async function main() {
-  setLoggingPrefix(jobName);
-  const startTime = startTimer(jobName);
-  if (!runNpmChecks()) {
-    return abortTimedJob(jobName, startTime);
+function pushBuildWorkflow() {
+  timedExecOrDie('gulp update-packages');
+  timedExecOrDie('gulp check-exact-versions');
+  timedExecOrDie('gulp lint');
+  timedExecOrDie('gulp prettify');
+  timedExecOrDie('gulp presubmit');
+  timedExecOrDie('gulp ava');
+  timedExecOrDie('gulp babel-plugin-tests');
+  timedExecOrDie('gulp caches-json');
+  timedExecOrDie('gulp dev-dashboard-tests');
+  timedExecOrDie('gulp check-renovate-config');
+  timedExecOrDie('gulp server-tests');
+  timedExecOrDie('gulp dep-check');
+  timedExecOrDie('gulp check-types');
+  timedExecOrDie('gulp check-sourcemaps');
+  timedExecOrDie('gulp performance-urls');
+}
+
+async function prBuildWorkflow() {
+  const buildTargets = determineBuildTargets();
+  await reportAllExpectedTests(buildTargets);
+  timedExecOrDie('gulp update-packages');
+
+  timedExecOrDie('gulp check-exact-versions');
+  timedExecOrDie('gulp lint');
+  timedExecOrDie('gulp prettify');
+  timedExecOrDie('gulp presubmit');
+  timedExecOrDie('gulp performance-urls');
+
+  if (buildTargets.has('AVA')) {
+    timedExecOrDie('gulp ava');
   }
 
-  if (!isPullRequestBuild()) {
-    timedExecOrDie('gulp update-packages');
-    timedExecOrDie('gulp check-exact-versions');
-    timedExecOrDie('gulp lint');
-    timedExecOrDie('gulp prettify');
-    timedExecOrDie('gulp presubmit');
-    timedExecOrDie('gulp ava');
+  if (buildTargets.has('BABEL_PLUGIN')) {
     timedExecOrDie('gulp babel-plugin-tests');
+  }
+
+  if (buildTargets.has('CACHES_JSON')) {
     timedExecOrDie('gulp caches-json');
+  }
+
+  // Check document links only for PR builds.
+  if (buildTargets.has('DOCS')) {
+    timedExecOrDie('gulp check-links --local_changes');
+  }
+
+  if (buildTargets.has('DEV_DASHBOARD')) {
     timedExecOrDie('gulp dev-dashboard-tests');
+  }
+
+  // Validate owners syntax only for PR builds.
+  if (buildTargets.has('OWNERS')) {
+    timedExecOrDie('gulp check-owners --local_changes');
+  }
+
+  if (buildTargets.has('RENOVATE_CONFIG')) {
     timedExecOrDie('gulp check-renovate-config');
+  }
+
+  if (buildTargets.has('SERVER')) {
     timedExecOrDie('gulp server-tests');
+  }
+
+  if (buildTargets.has('RUNTIME')) {
     timedExecOrDie('gulp dep-check');
     timedExecOrDie('gulp check-types');
     timedExecOrDie('gulp check-sourcemaps');
-    timedExecOrDie('gulp performance-urls');
-  } else {
-    printChangeSummary();
-    const buildTargets = determineBuildTargets();
-    await reportAllExpectedTests(buildTargets);
-    timedExecOrDie('gulp update-packages');
-
-    timedExecOrDie('gulp check-exact-versions');
-    timedExecOrDie('gulp lint');
-    timedExecOrDie('gulp prettify');
-    timedExecOrDie('gulp presubmit');
-    timedExecOrDie('gulp performance-urls');
-
-    if (buildTargets.has('AVA')) {
-      timedExecOrDie('gulp ava');
-    }
-
-    if (buildTargets.has('BABEL_PLUGIN')) {
-      timedExecOrDie('gulp babel-plugin-tests');
-    }
-
-    if (buildTargets.has('CACHES_JSON')) {
-      timedExecOrDie('gulp caches-json');
-    }
-
-    // Check document links only for PR builds.
-    if (buildTargets.has('DOCS')) {
-      timedExecOrDie('gulp check-links --local_changes');
-    }
-
-    if (buildTargets.has('DEV_DASHBOARD')) {
-      timedExecOrDie('gulp dev-dashboard-tests');
-    }
-
-    // Validate owners syntax only for PR builds.
-    if (buildTargets.has('OWNERS')) {
-      timedExecOrDie('gulp check-owners --local_changes');
-    }
-
-    if (buildTargets.has('RENOVATE_CONFIG')) {
-      timedExecOrDie('gulp check-renovate-config');
-    }
-
-    if (buildTargets.has('SERVER')) {
-      timedExecOrDie('gulp server-tests');
-    }
-
-    if (buildTargets.has('RUNTIME')) {
-      timedExecOrDie('gulp dep-check');
-      timedExecOrDie('gulp check-types');
-      timedExecOrDie('gulp check-sourcemaps');
-    }
   }
-
-  stopTimer(jobName, startTime);
 }
 
-main();
+runCiJob(jobName, pushBuildWorkflow, prBuildWorkflow);
